miniflow_extract: Push ct_tp_{src,dst} for later frags.
Ales reported that ct_tp_{src,dst} matches are working only for the
first frag for the userspace datapath, whereas they are always working
for later frags in the case of kernel datapath.

The ipf propagates the info in packets metadata, but
miniflow_extract() has no handling for them.

Fix it by pushing the relevant fields in the miniflow.
tp_{src,dst} are not set for later frags, so fill them with padding as
ct_tp_{src,dst} are not aligned:

struct flow {
    [...]
    ovs_be16                   tp_src;               /*   656     2 */
    ovs_be16                   tp_dst;               /*   658     2 */
    ovs_be16                   ct_tp_src;            /*   660     2 */
    ovs_be16                   ct_tp_dst;            /*   662     2 */
    [...]
}

The patch also includes two tests to exercise the behavior.

Reported-at: https://issues.redhat.com/browse/FDP-124
Fixes: 4ea96698f667 ("Userspace datapath: Add fragmentation handling.")
Signed-off-by: Paolo Valerio <pvalerio@redhat.com>
---
v2:
- Rearranged fragmented IPv6 packets in the test (Ilya)
- Added tshark dissection output for all packets (Ilya)
- while at it, changed the IPv4/UDP packet using a payload with
  repetitions to make it smaller
- updated IPv6 fragmentation test rules to make them more robust.
  Readded `-s 1600` option to ping6 as it must have disappeared
  before posting v1 for some reason.